@@ -2,7 +2,7 @@ package com.interstellarstudios.weathershare;
 
 import android.graphics.Color;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
 import gr.net.maroulis.library.EasySplashScreen;
 
